requirements.txt
# Microsoft 365 Email Automation System Dependencies

# Microsoft Authentication Library
msal>=1.24.0

# HTTP client for Microsoft Graph API
requests>=2.31.0
aiohttp>=3.8.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Template engine
jinja2>=3.1.0

# Scheduler
apscheduler>=3.10.0

# Environment configuration
python-dotenv>=1.0.0

# Email validation
email-validator>=2.0.0

# Logging and utilities
structlog>=23.0.0

# Development dependencies (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0